refactor: 캘린더 구조 정리
