+++ src/features/Calendar/components/CalendarView/day/renderGeometry.ts
