+++ src/features/Calendar/components/CalendarView/eventStyles.ts
