+++ src/features/Calendar/components/CalendarView/month/MonthEvent.tsx
@@ -4,10 +4,10 @@ import type { EventProps } from 'react-big-calendar'
 import { createPortal } from 'react-dom'
 
 import People from '@/assets/icons/people.svg?react'
+import type { CalendarEvent } from '@/shared/types/calendar/types'
 
-import { getColorPalette } from '../../utils/colorPalette'
-import type { CalendarEvent } from '../CustomView/CustomDayView'
-import * as S from './CustomEvent.style'
+import { getColorPalette } from '../../../utils/colorPalette'
+import * as S from '../eventStyles'
 
 const formatTimeRange = (event: CalendarEvent) => {
   if (event.isAllDay) {
