+++ src/features/Calendar/components/CalendarView/month/MonthShowMore.tsx
@@ -2,7 +2,7 @@ import type { ShowMoreProps } from 'react-big-calendar'
 
 import type { CalendarEvent } from '@/shared/types/calendar/types'
 
-import * as S from './CustomEvent.style'
+import * as S from '../eventStyles'
 
 const CustomMonthShowMore = ({ count }: ShowMoreProps<CalendarEvent>) => (
   <S.MonthShowMore>{`+${count}개`}</S.MonthShowMore>
