+++ src/features/Calendar/components/CalendarView/month/index.ts
@@ -1,0 +1,2 @@
+export { default as CustomMonthEvent } from './MonthEvent'
+export { default as CustomMonthShowMore } from './MonthShowMore'
