+++ src/features/Calendar/components/CalendarView/week/weekLayout.ts
@@ -7,7 +7,7 @@ import {
   eventCoversDate,
   getEventOccurrenceKey,
   isDateOnlyString,
-} from './helpers/dayViewHelpers'
+} from '../../../utils/helpers/dayViewHelpers'
 
 export const KOREAN_WEEKDAYS = ['일', '월', '화', '수', '목', '금', '토'] as const
