--- src/features/Calendar/components/CustomEvent/index.tsx
@@ -1,3 +1,0 @@
-export { default as CustomMonthEvent } from './CustomMonthEvent'
-export { default as CustomMonthShowMore } from './CustomMonthShowMore'
-export { default as CustomWeekEvent } from './CustomWeekEvent'
